requirements.txt
openai==0.27.0
python-dotenv==1.0.0
discord.py==2.2.2
tiktoken==0.3.0
